backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
httpx
openai
elevenlabs
python-multipart
pyjwt
